==> data/code/99_folklore.R <==
# Load packages
source("00_load_packages.R")



############### 1. Folklore Dataset ##############

# Load D-Place EA
load(here("data". "output", ""))

# Folklore-based measures
mturk_ea <- read_dta(here("data", "raw", "folklore", "Original_Files",
                          "MTurk_EA_groups.dta"))
